fix: change unsigned int to size_t

+++ more_malloc_free/2-calloc.c
@@ -11,7 +11,7 @@
 
 void *_calloc(unsigned int nmemb, unsigned int size)
 {
-	unsigned int i, total;
+	size_t i, total;
 	void *ptr;
 	unsigned char *p;
